Fixed formatting problems and now able to show details.
Changed the way formatting works: it groups activities together, and optimises
the process of gathering information by visiting every file only once.

With this change, it becomes easier to store the detailed properties about each
activity during this process. So it is now possible to view every activity in
detail.